Saner pasting rules
You can have rolling buffers for delete ("1-"9) but not for yank
(only "0), but by default you paste from "" which will have both yank
and delete but! when you delete the yank doesn't get rolled so when you
do a new yank the prev yank gets lost but when you do a new delete the
prev delete is kept only! if the delete contained a new line character
else it'll also be lost (from "-) akin to yank.

Amazing.